+++ src/adapter.js
@@ -114,6 +114,7 @@
     // Suppress searching if it returns true.
     _skipSearch: function (clickEvent) {
       switch (clickEvent.keyCode) {
+        case 9:  // TAB
         case 13: // ENTER
         case 40: // DOWN
         case 38: // UP